backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic[email]==2.10.3
pydantic-settings==2.6.1
supabase==2.10.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9
weasyprint==62.3
jinja2==3.1.4
python-multipart==0.0.9
httpx==0.27.0
